python-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai-whisper==20231117
transformers==4.35.0
torch==2.1.0
numpy==1.24.3
spacy==3.7.2
pydub==0.25.1
moviepy==1.0.3